# fix threads per block in inner persistent heuristics