--- a/csrc/scheduler/normalization.cpp
+++ b/csrc/scheduler/normalization.cpp
@@ -598,10 +598,9 @@ std::shared_ptr<ReductionParams> innerPersistentHeuristic(
     // than five percent.
     if (batches_per_block_inner_reduction >
         batches_per_block_inner_reduction_max) {
-      batches_per_block_inner_reduction /= 2;
-      bdimx = ceilDiv(
-          inner_most_dimension_numel,
-          inner_reduction_unroll_factor * batches_per_block_inner_reduction);
+      bdimx *= 2;
+      batches_per_block_inner_reduction = ceilDiv(
+          inner_most_dimension_numel, inner_reduction_unroll_factor * bdimx);
       continue;
     }
 
@@ -692,7 +691,6 @@ std::shared_ptr<ReductionParams> innerPersistentHeuristic(
     const int64_t blocks_per_sm_estimated =
         getThreadsPerSMGivenRegPerThread(estimated_register_count) /
         threads_per_block;
-
     // only allow adjust to 90% of estimated_register_count to avoid too much
     // spills. initially we used 80%, however, the drop from 160 to 128 leads to
     // too much spills in Layer Norm with fused ops, see