1. In the test of softmax found an error with hidden size of 22024. It leads to threads per block larger than 1024 due to `ceilDiv`. In this PR, we update bdimx first then update batches per block.
2. Extended `FusionMagicSchedulerSoftmax_CUDA` to allow more test cases if `const bool test_all = true;` by default it is set to `false`.