@liqiangxl 2023-06-03
Perf changed only on one case with a feature size of 5140. Looks like all other cases are dividable.
Main branch (old), batch = 5, effective threads = 5140 / 4 / 5 = 257, launched = 288
```
ptxas info    : 307 bytes gmem
ptxas info    : Compiling entry function '_ZN11CudaCodeGen7kernel5ENS_6TensorINS_6__halfELi1ELi1EEENS0_IS1_Li2ELi2EEES2_S2_S2_S3_' for 'sm_80'
ptxas info    : Function properties for _ZN11CudaCodeGen7kernel5ENS_6TensorINS_6__halfELi1ELi1EEENS0_IS1_Li2ELi2EEES2_S2_S2_S3_
ptxas         .     0 bytes stack frame, 0 bytes spill stores, 0 bytes spill loads
ptxas info    : Used 63 registers, 16 bytes smem, 464 bytes cmem[0], 8 bytes cmem[2]

blocks_per_sm= 3, warps_per_sm= 27, occupancy= 42.19%
NvFuserScheduler_LayerNormFusedOp_fp16___GRAPH/NvFuserScheduler_LayerNormFusedOp_fp16/8192/5140/manual_time         241 us          316 us           58 bytes_per_second=651.193G/s Red On Fastest Dim // Persistent Kernel //  // Iteration Domain:  // Inner Reduction Domain: cross block reduction / pad to warp / persistent batch - 5 / vectorize / factor 4/Launch_Parameters[block(1/1/288)/grid(1/1/8192)/1152]
```
With this change (new): batch = 6, effective threads = 5140 / 4 / 6 = 215, launched = 224
```
ptxas info    : 307 bytes gmem
ptxas info    : Compiling entry function '_ZN11CudaCodeGen7kernel5ENS_6TensorINS_6__halfELi1ELi1EEENS0_IS1_Li2ELi2EEES2_S2_S2_S3_' for 'sm_80'
ptxas info    : Function properties for _ZN11CudaCodeGen7kernel5ENS_6TensorINS_6__halfELi1ELi1EEENS0_IS1_Li2ELi2EEES2_S2_S2_S3_
ptxas         .     0 bytes stack frame, 0 bytes spill stores, 0 bytes spill loads
ptxas info    : Used 71 registers, 16 bytes smem, 464 bytes cmem[0], 8 bytes cmem[2]
blocks_per_sm= 4, warps_per_sm= 28, occupancy= 43.75%
NvFuserScheduler_LayerNormFusedOp_fp16___GRAPH/NvFuserScheduler_LayerNormFusedOp_fp16/8192/5140/manual_time         210 us          284 us           67 bytes_per_second=746.823G/s Red On Fastest Dim // Persistent Kernel //  // Iteration Domain:  // Inner Reduction Domain: cross block reduction / pad to warp / persistent batch - 6 / vectorize / factor 4/Launch_Parameters[block(1/1/224)/grid(1/1/8192)/896]
```

![image](https://github.com/NVIDIA/Fuser/assets/116412316/120a17f3-b7f0-45f6-89ee-8c6f4347194f)