@naoyam 2023-06-03
>Looks like all other cases are dividable.

Are you sure? Can you please confirm it?