@liqiangxl 2023-06-05
> 



> Are you sure? Can you please confirm it?

You are right. There are also other cases but there is no noticable change in perf.
![image](https://github.com/NVIDIA/Fuser/assets/116412316/b42e06d3-02de-4c14-9908-a225e3af13f2)